# Benusy/automation/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Influencer Promotion Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tasks/": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List Tasks",
        "operationId": "list_tasks_api_v1_tasks__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskRead"
                  },
                  "type": "array",
                  "title": "Response List Tasks Api V1 Tasks  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task",
        "operationId": "get_task_api_v1_tasks__task_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/accept": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Accept Task",
        "operationId": "accept_task_api_v1_tasks__task_id__accept_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/tasks": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Tasks",
        "operationId": "list_tasks_api_v1_admin_tasks_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TaskStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskRead"
                  },
                  "title": "Response List Tasks Api V1 Admin Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Task",
        "operationId": "create_task_api_v1_admin_tasks_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/tasks/{task_id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Update Task",
        "operationId": "update_task_api_v1_admin_tasks__task_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignmentRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/AssignmentStatus"
          },
          "post_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Post Link"
          },
          "reject_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reject Reason"
          },
          "metric_sync_status": {
            "$ref": "#/components/schemas/MetricSyncStatus"
          },
          "last_sync_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync Error"
          },
          "revenue": {
            "type": "number",
            "title": "Revenue"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "last_synced_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Synced At"
          },
          "task": {
            "$ref": "#/components/schemas/TaskRead"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricRead"
            },
            "type": "array",
            "title": "Metrics"
          },
          "manual_metric_submissions": {
            "items": {
              "$ref": "#/components/schemas/ManualMetricSubmissionRead"
            },
            "type": "array",
            "title": "Manual Metric Submissions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "metric_sync_status",
          "revenue",
          "created_at",
          "updated_at",
          "task"
        ],
        "title": "AssignmentRead"
      },
...
      "AssignmentStatus": {
        "type": "string",
        "enum": [
          "accepted",
          "in_review",
          "rejected",
          "completed",
          "cancelled"
        ],
        "title": "AssignmentStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualMetricReviewStatus": {
        "type": "string",
        "enum": [
          "pending",
          "approved",
          "rejected"
        ],
        "title": "ManualMetricReviewStatus"
      },
      "ManualMetricSubmissionRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "assignment_id": {
            "type": "integer",
            "title": "Assignment Id"
          },
          "likes": {
            "type": "integer",
            "title": "Likes"
          },
          "favorites": {
            "type": "integer",
            "title": "Favorites"
          },
          "shares": {
            "type": "integer",
            "title": "Shares"
          },
          "views": {
            "type": "integer",
            "title": "Views"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "review_status": {
            "$ref": "#/components/schemas/ManualMetricReviewStatus"
          },
          "review_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Reason"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "assignment_id",
          "likes",
          "favorites",
          "shares",
          "views",
          "review_status",
          "submitted_at"
        ],
        "title": "ManualMetricSubmissionRead"
      },
...
      "MetricRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "likes": {
            "type": "integer",
            "title": "Likes"
          },
          "favorites": {
            "type": "integer",
            "title": "Favorites"
          },
          "shares": {
            "type": "integer",
            "title": "Shares"
          },
          "views": {
            "type": "integer",
            "title": "Views"
          },
          "source": {
            "$ref": "#/components/schemas/MetricSource"
          }
        },
        "type": "object",
        "required": [
          "id",
          "timestamp",
          "likes",
          "favorites",
          "shares",
          "views",
          "source"
        ],
        "title": "MetricRead"
      },
      "MetricSource": {
        "type": "string",
        "enum": [
          "auto",
          "manual"
        ],
        "title": "MetricSource"
      },
      "MetricSyncStatus": {
        "type": "string",
        "enum": [
          "normal",
          "manual_required",
          "manual_pending_review",
          "manual_approved",
          "manual_rejected"
        ],
        "title": "MetricSyncStatus"
      },
...
      "TaskCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "base_reward": {
            "type": "number",
            "minimum": 0.0,
            "title": "Base Reward"
          },
          "accept_limit": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Accept Limit"
          },
          "instructions": {
            "type": "string",
            "title": "Instructions"
          },
          "attachments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Attachments"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "default": "draft"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "platform",
          "base_reward",
          "instructions"
        ],
        "title": "TaskCreate"
      },
...
      "TaskRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "base_reward": {
            "type": "number",
            "title": "Base Reward"
          },
          "accept_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accept Limit"
          },
          "instructions": {
            "type": "string",
            "title": "Instructions"
          },
          "attachments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Attachments"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "accepted_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted Count"
          },
          "remaining_slots": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining Slots"
          },
          "is_full": {
            "type": "boolean",
            "title": "Is Full",
            "default": false
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "platform",
          "base_reward",
          "instructions",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "TaskRead"
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published",
          "cancelled"
        ],
        "title": "TaskStatus"
      },
      "TaskUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "platform": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform"
          },
          "base_reward": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Reward"
          },
          "accept_limit": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Accept Limit"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "attachments": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachments"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "TaskUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        }
      }
    }
  }
}